Refactorizat functionalitati Cupid in Watcher
