+++ celluloid/src/main/java/com/celluloid/GameOfLife.java
@@ -9,14 +9,13 @@ import java.util.ArrayList;
 public class GameOfLife {
     private final FoodPool foodPool = new FoodPool();
     private final Watcher watcher = new Watcher(foodPool);
-    private final Cupid cupid = new Cupid();
 
     private final ArrayList<SexualCell> sexualCells = new ArrayList<>();
     private final ArrayList<AsexualCell> asexualCells = new ArrayList<>();
 
     public GameOfLife() {
         for (int i = 0; i < Config.SEXUAL_CELLS_COUNT; i++) {
-            sexualCells.add(new SexualCell("sexual_cell_" + i, foodPool, watcher, cupid));
+            sexualCells.add(new SexualCell("sexual_cell_" + i, foodPool, watcher));
         }
 
         for (int i = 0; i < Config.ASEXUAL_CELLS_COUNT; i++) {
@@ -27,7 +26,6 @@ public class GameOfLife {
     @PostConstruct
     public void startSimulation() {
         watcher.start();
-        cupid.start();
 
         for (var cell : sexualCells) {
             Thread thread = new Thread(cell);
